Make sure signature stays the same in all nginx builds
When a dyn module is loaded the signature of the module is compared to
the one of the nginx binary.

dyn modules are build from nginx-full, so in order to make them loadable
in other flavors we need to make sure all the binaries share the same
signature.

The signature is a vector computed from 33+3 parameters (src/core/ngx_module.h)
With this commit we align all nginx flavors to same signature vector.