Reuse the same build flags for the perl module
Using the same flags broke the build since -shared was not positioned
after pie/pic flags. Changing the perl LDDFLAGS patch and moving
Config{lddflags} to the end solved that issue.

As a general note, the injected -fPIC in debian/rules is not needed
since dpkg 1.18.13. We inject -fPIC so that it overrides previous -PIE
flag, but the latest dpkg automatically strips pie flags when building
shared libraries (man dpkg-buildflags).

We temporary keep the -fPIC injection for wheezy/jessie backports and
downstream Ubuntu builds.