Automate modules watch & upgrade process
This patch introduces the debian/ngxmod script and some helper
files which can be used to automate the modules watch & upgrade process.

The only subcommand for now is `uscan` and can be called as:
  $ debian/ngxmod uscan

The uscan subcommand runs uscan for each nginx module listed in
the newly deb822 formatted debian/modules/control using the watchfiles
in debian/modules/watch.

If a new version is available, it will ask you if you want to upgrade.
If you agree, it will download the tarball, place it in the nginx
source's parent directory and upgrade the module's source.

After the upgrade, it will remove the files listed in the
Files-Excluded module field and finally it will commit the changes.

Closes: #869499